PriorityQueueとNodeのDistanceをUtilsに移動
